web/code/requirements.txt
astroid==1.4.8
boltons==16.5.0
candybar==0.1.5
Django==1.9.9
django-address==0.1.5
django-autoslug==1.9.3
django-bootstrap3==7.0.1
django-cleanup==0.4.2
django-extensions==1.7.1
django-grappelli==2.8.1
django-mptt-nomagic==0.2.3
docopt==0.6.2
fake-factory==0.6.0
graphviz==0.4.10
isort==4.2.5
jedi==0.9.0
lazy-object-proxy==1.2.2
mccabe==0.5.2
pep8==1.7.0
Pillow==3.3.0
prompt-toolkit==1.0.5
psycopg2==2.6.2
ptpython==0.35
pycountry==1.20
Pygments==2.1.3
pygraphviz==1.3.1
pylint==1.6.4
python-dateutil==2.5.3
six==1.10.0
wcwidth==0.1.7
Werkzeug==0.11.10
wrapt==1.10.8
